Rename EditConfiguration -> Configuration
