+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activity.editconfiguration.EditConfigurationActivity"></activity>
+        <activity android:name=".activity.configuration.ConfigurationActivity"></activity>
     </application>
 
 </manifest>
