+++ app/src/main/java/com/cannawen/dota2timer/activity/configuration/ConfigurationActivity.java
@@ -1,4 +1,4 @@
-package com.cannawen.dota2timer.activity.editconfiguration;
+package com.cannawen.dota2timer.activity.configuration;
 
 import android.app.Activity;
 import android.content.Context;
@@ -13,12 +13,12 @@ import com.cannawen.dota2timer.configuration.Configuration;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class EditConfigurationActivity extends Activity {
+public class ConfigurationActivity extends Activity {
 
     private static final String INTENT_CONFIGURATION_KEY = "INTENT_CONFIGURATION_KEY";
 
     static public Intent createActivityIntent(Context context, Configuration configuration) {
-        Intent intent = new Intent(context, EditConfigurationActivity.class);
+        Intent intent = new Intent(context, ConfigurationActivity.class);
         intent.putExtra(INTENT_CONFIGURATION_KEY, configuration);
         return intent;
     }
@@ -27,13 +27,13 @@ public class EditConfigurationActivity extends Activity {
         return (Configuration) intent.getSerializableExtra(INTENT_CONFIGURATION_KEY);
     }
 
-    @BindView(R.id.activity_edit_configuration_recycler_view)
+    @BindView(R.id.activity_configuration_recycler_view)
     RecyclerView recyclerView;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_edit_configuration);
+        setContentView(R.layout.activity_configuration);
         ButterKnife.bind(this);
 
         recyclerView.setLayoutManager(new LinearLayoutManager(getApplicationContext(), LinearLayoutManager.VERTICAL, false));
